dev-generator: create function to create resource using the generator
After creating the resource with the generator, it automatically creates a file for it. As it seems that we cannot specify parameters for the file creation, we get the file that has been created for the resource, and we update its properties using the same parameters used for the resource.
For the test, probably, both resource and file objects will be needed, so, we return both in an array.