issue #6: create function to create the zip file name basing on the current timestamp + milliseconds
Is necessary to give each zip file an almost unique name, to avoid possible collisions of more than one person generating the zip for the same course. And, using the datetime with milliseconds of the time of the request, come on, is almost impossible to make that happen :) The format of the name is in human readable format instead of pure numbers, to make easier to the people downloading the zip to identify, classify or whatever they want to do with them; a string of milliseconds does not tell anything.