issue #7: use a session variable to check if the user has had access before trying the download
Otherwise, if a user goes directly to the download.php page, it could download any file, specifying the path in the parameter. And, in download.php, there's no chance to check if the user is enrolled in the course, because the courseid is not passed. And passing it would be a silly decission: a course id and a path to a zip file that not belongs to it can be passed.