issue #6: modify "create_zip_name" function to use the combination of user and course for the name
Is necessary to give each zip file an unique name, to avoid possible collisions of more than one person generating the zip for the same course. And, generating an unique zip file for the combination of user and course, every time an user tries to generate the zip for the course, the previous will be overwritten (if exists), so, the files won't be accumulating occupying useful space.