dev-folder: code test of "add_resources_to_directory", but mark to skip
This is not testeable since a way to generate files in folders natively, with the data generator, is found. Because, the 'create_resource_in_folder' updates manually the files' table, in order to make the query of 'get_db_records' return the files in folders, but then the file API is not able to retrieve the files with those modified values.