issue #14: change file and directory names to ASCII (and transliterate)
file_put_contents function is not able to properly encode the file names in UTF-8. So, the best option is to use ASCII. Sorry for those chinese, japanese, etc. Moodle users. :(
The '//TRANSLIT' suffix activates the transliteration, i.e., when a character can't be represented in the target charset, it can be approximated through one or several similarly looking characters. Now, the cleanup of forbidden characters is made after the transliteration, because this process may create illegal characters for those non-transliterable characters.